[fixed]:(dashboard,contact):update user dashboard for small device and update contact page
